LP_CnC_OT now works
Pre-generated primes for the Schnorr group. But the subgroup is still
chosen at random each time. Also incorporated into the OT_PWV, so we now
actually use a  group for which the DDH assumption might hold.

Also a bit of re-structuring of code for the sake of my continued
sanity.